Better support GHC 7.10 by loosening bound on directory
directory is a boot library, meaning that it is one of the libraries
used to build GHC itself. Any library that depends on ghc-the-library
must use the exact boot library versions that that ghc-the-compiler
version was built with. The GHC 7.10 series was built with directory
1.2.2.0, so to best support users on GHC 7.10, we should support
their GHC boot library versions of dependencies. Otherwise we are
incompatible with them if they depend on ghc-the-library

listDirectory was not in 1.2.2.0 so I've replaced it with
getDirectoryContents. The difference between them is that
listDirectory filters its output. But every usage of listDirectory
in this package is followed by a stricter filter than the one in
listDirectory, so getDirectoryContents should be fine.